% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt_run.R
\name{bbt_run}
\alias{bbt_run}
\title{BBT rating algorithm}
\usage{
bbt_run(formula, data, r, rd, sig, weight)
}
\arguments{
\item{formula}{formula specifying model. BBT algorithm allows only player ranking parameter and should be specified by following manner: 
`rank | id ~ name`. Names in formula are unrestricted, but model structure remains the same where:
\enumerate{
 \item rank alternative/participant position in experiment/event. Order of comparison
 \item id event identifier
 \item name of alternative/participant
}}

\item{data}{data.frame which contains columns specified in formula, and optionaly columns defined by `sig`, `weight` or `date`.}

\item{r}{named vector of initial rating estimates. In there is no assumption, initial ratings recomended to be r=25 Names of vector should correspond with team_name label.}

\item{rd}{named vector of initial rating deviation estimates. In there is no assumption, initial ratings recomended to be r=25/3 Names of vector should correspond with team_name label.}

\item{sig}{name of column in `data` containing rating volatility. Rating volitality is a value which multiplies prior `rd`. If `sig > 1` then prior `rd` increases, making estimate of `r` more uncertain.}

\item{weight}{name of column in `data` containing weights. Weights multiplies step update increasing/decreasing step impact on parameters estimates}

\item{date}{name of column in `data` containing date. Doesn't affect estimation process. If specified, charts displays estimates changes in time instead of by observation `id`}
}
\description{
BBT rating algorithm
Wrapper arround `bbt` update algorithm. Wrapper allows user to simplify calculation providing only data and initial parameters assumptions
}
