% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bdl}
\alias{bdl}
\title{Bayesian Bradley-Terry model for single game}
\usage{
bdl(team_name, rank, R, X, RD, map)
}
\arguments{
\item{team_name}{player/team names.}

\item{X}{Matrix of coefficients (ratings).}

\item{H}{Matrix of player specifics.}

\item{S}{Matrix of coefficients deviations}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates Glicko ratings based on Bayesian Bradley Terry model.
}
\details{
Algorithm based on 'A Bayesian Approximation Method for Online Ranking' by Ruby C. Weng and Chih-Jen Lin
}
