% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTermMatrix.R
\name{createTermMatrix}
\alias{createTermMatrix}
\title{Creates Full Term Matrix}
\usage{
createTermMatrix(formula, data)
}
\arguments{
\item{formula}{Specify model by formula.}

\item{data}{data.frame}
}
\description{
Creates Full Term Matrix
}
