% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlr1}
\alias{dlr1}
\title{Dynamic Logit Regression}
\usage{
dlr1(team_name, rank, R, X, RD)
}
\arguments{
\item{team_name}{player/team names.}

\item{R}{Matrix of coefficients (ratings).}

\item{X}{Matrix of player specifics.}

\item{RD}{Matrix of coefficients deviations}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates ratings using extended Kalman Filter.
}
\details{
Algorithm based on 'A Bayesian Approximation Method for Online Ranking' by Ruby C. Weng and Chih-Jen Lin
}
