% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_dlr.R
\name{dlr_run}
\alias{dlr_run}
\title{BDL rating algorithm}
\usage{
dlr_run(formula, data, r, rd)
}
\arguments{
\item{formula}{formula specifying model.}

\item{data}{data.frame which contains three columns: #' \enumerate{
  \item id event identifier
  \item team_name label of event participant -  name corresponding with parameters names
  \item rank position of team_name in event
}}

\item{r}{named vector of initial rating estimates. In there is no assumption, initial ratings should be r=25 Names of vector should correspond with team_name label.}

\item{rd}{named vector of initial rating deviation estimates. In there is no assumption, initial ratings should be r=25/3 Names of vector should correspond with team_name label.}
}
\description{
BDL rating algorithm
Wrapper arround `bdl` update algorithm. Wrapper allows user to simplify calculation providing only data and initial parameters assumptions
}
