% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dynamicBT}
\alias{dynamicBT}
\title{Glicko rating for single game}
\usage{
dynamicBT(rank, mi_ij, sig_ij, kappa = 0.95, gamma = 1)
}
\arguments{
\item{rank}{classification of the event.}

\item{teams}{name of event participants.}

\item{days}{days after previous match - indicator multiplying uncertainty of expectations.}

\item{r}{ratings of participants.}

\item{rd}{rating deviations of participants.}

\item{init_r}{initial rating for new competitors (contains NA). Default = 1500}

\item{init_rd}{initial rating deviations for new competitors. Default = 350}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates Glicko rating for single game input
}
