% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elo}
\alias{elo}
\title{FIDE rating for single game}
\usage{
elo(teams, rank, r, K = 32L, init_r = 1500, init_rd = 350)
}
\arguments{
\item{teams}{name of event participants.}

\item{rank}{classification of the event.}

\item{r}{ratings of participants.}

\item{init_r}{initial rating for new competitors (contains NA). Default = 1500}

\item{init_rd}{initial rating deviations for new competitors. Default = 350}

\item{days}{days after previous match - indicator multiplying uncertainty of expectations.}

\item{rd}{rating deviations of participants.}
}
\description{
FIDE Rating Regulations effective from 1 July 2014
}
\examples{
elo(
 teams = c( "A", "B", "C", "D", "E", "F" ), 
 rank  = c( 3, 2, 3, 5, 6, 1), 
 r     = c( 1613, 1609, 1477, 1388, 1586, 1720 ) , 
 init_r  = 1500,
 init_rd = 100
)
}
