% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factor2dummy}
\alias{factor2dummy}
\title{Dummy matrix from factor}
\usage{
factor2dummy(factor)
}
\arguments{
\item{factor}{to be transormed into dummy matrix}
}
\description{
Dummy matrix from factor
}
