% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glicko}
\alias{glicko}
\title{Glicko rating for single game}
\usage{
glicko(teams, rank, days, names, r, rd, init_r = 1500, init_rd = 350)
}
\arguments{
\item{r}{ratings of game participants.}

\item{RD}{rating deviations.}
}
\description{
Calculates Glicko rating for single game input
}
\examples{
 teams <- c("A","B","C","D")
 rank  <- c(3 , 4 , 1, 2)

 names <- c("A","B","E","F","C","D","E")
 r     <- c(1500, 1400, 1500, 1500, 1550, 1700) 
 rd    <- c(200,  30,   300,  300,  100,  300)
 glicko(
   teams = teams, 
   rank  = rank, 
   days  = rep(1,length(rank)),
   names = names, 
   r     = r, 
   rd    = rd,
   init_r  = 1500,
   init_rd = 100)
}

