% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glicko2}
\alias{glicko2}
\title{Glicko2 rating for single game}
\usage{
glicko2(team_name, rank, r, rd, sig, time = as.numeric(c(0)), tau = 0.5,
  init_r = 1500, init_rd = 350)
}
\arguments{
\item{team_name}{name of event participants.}

\item{rank}{classification of the event.}

\item{r}{ratings of participants.}

\item{rd}{rating deviations of participants.}

\item{time}{time after previous match - indicator multiplying uncertainty of expectations.}

\item{init_r}{initial rating for new competitors (contains NA). Default = 1500}

\item{init_rd}{initial rating deviations for new competitors. Default = 350}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates Glicko2 rating for single game input
}
\examples{
glicko2(
 team_name = c( "A", "B", "C", "D" ), 
 rank  = c( 3, 4, 1, 2 ), 
 time  = c( 0, 0, 0, 0),
 r     = c( 1500, 1400, 1550, 1700 ) , 
 rd    = c( 200,  30,   100,  300 ),
 sig   = c( .06, .06, .05, .07),
 tau   = .5,
 init_r  = 1500,
 init_rd = 100
)
}
