% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gpheats}
\alias{gpheats}
\title{Heat results of Speedway Grand-Prix}
\format{
A data frame with >19000 rows and 11 variables:
\describe{
\itemize{
\item \code{id} - event identifier
\item \code{season} - year of Grand-Prix, 1995-now
\item \code{date} - date of turnament
\item \code{round} - round in season
\item \code{name} - Turnament name
\item \code{heat} - heat number, 1-23
\item \code{field} - number of gate, 1-4
\item \code{rider} - rider name, string
\item \code{points} - paints gained, integer
\item \code{position} - position at finish line, string
\item \code{rank} - rank at finish line, integer
}
}
}
\source{
internal
}
\description{
Actual dataset containing heats results of all Speedway Grand-Prix turnaments
\code{gpheats}.
}
