% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpsquads}
\alias{gpsquads}
\title{Turnament results of Speedway Grand-Prix}
\format{A data frame with >4000 rows and 9 variables:
\describe{
  \item{event_id}{event identifier}
  \item{season}{year of Grand-Prix, 1995-now}
  \item{date}{date of turnament}
  \item{round}{round in season}
  \item{name}{Turnament name}
  \item{place}{stadium of event}
  \item{rider_name}{points, 1-6}
  \item{classification}{classification after an event}
}}
\source{
\url{juicystat.com}
}
\usage{
gpsquads
}
\description{
Actual dataset containing turnament results of all Speedway Grand-Prix events
gpsquads
}
\keyword{datasets}

