% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_logit.R
\name{kalman_logit}
\alias{kalman_logit}
\title{Dynamic Logistic Regression}
\usage{
kalman_logit(x, z, H, P, Bu, pa)
}
\description{
Dynamic Logistic Regression
R implementation of 'Dynamic Logistic Regression' William D. Penny and Stephen J. ROberts
}
