% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KF}
\alias{KF}
\title{Kalman filter for single game}
\usage{
KF(x, F, B, u, z, H)
}
\arguments{
\item{teams}{name of event participants.}

\item{rank}{classification of the event.}

\item{days}{days after previous match - indicator multiplying uncertainty of expectations.}

\item{r}{ratings of participants.}

\item{rd}{rating deviations of participants.}

\item{init_r}{initial rating for new competitors (contains NA). Default = 1500}

\item{init_rd}{initial rating deviations for new competitors. Default = 350}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates score by smoothing
}
\examples{
KF(
 x    = c( 1500, 1400, 1550, 1700 ) , 
 F    = diag(4),
 B    = matrix(c( 200,  30,   100,  300 ),2),
 u    = c( 1, 1, 1, 1),
 z    = c(0,1,0,0),
 H    = matrix(c( .06, .06, .05, .07),2)
)
}

