% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{plheats}
\alias{plheats}
\title{Heat results from Polish Speedway League}
\format{
A data frame with >170000 rows and 11 variables:
\describe{
  \item{id}{event identifier}
  \item{season}{year of Grand-Prix, 2004-2019}
  \item{season}{year of Grand-Prix, 2004-2019}
  \item{date}{date of turnament}
  \item{round}{round in season}
  \item{name}{Turnament name}
  \item{heat}{heat number, 1-15}
  \item{field}{number of gate, 1-4}
  \item{rider}{rider name, string}
  \item{points}{paints gained, integer}
  \item{position}{position at finish line, string}
  \item{rank}{rank at finish line, integer}
}
}
\source{
internal
}
\description{
Dataset containing heats results from Polish Speedway League
\code{plheats}.
}
