% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot.rating}
\alias{plot.rating}
\title{Plot rating object}
\usage{
\method{plot}{rating}(x, n = 10, teams, ...)
}
\arguments{
\item{x}{of class rating}

\item{n}{number of players to be plotted}

\item{...}{optional arguments}

\item{players}{optional vector with names of the contestants to plot 
their evolution in time.}
}
\description{
Plot rating object
}
