% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot_sport}
\alias{plot_sport}
\title{Plot sport object}
\usage{
plot_sport(object, n = 10, ...)
}
\arguments{
\item{object}{of class sport}

\item{n}{number of players to be plotted}

\item{...}{optional arguments}
}
\description{
Plot sport object
}
