% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{plsquads}
\alias{plsquads}
\title{Results from Polish Speedway League}
\format{
A data frame with >4000 rows and 9 variables:
\describe{
  \item{id}{event identifier}
  \item{season}{year of Grand-Prix, 2004-2019}
  \item{date}{date of turnament}
  \item{place}{stadium of event}
  \item{round}{round in season}
  \item{name}{Turnament name}
  \item{rider}{rider names, 1-6}
  \item{team_name}{rider names, 1-6}
  \item{points}{points gained, integer}
  \item{classification}{classification after an event}
}
}
\source{
internal
}
\description{
Dataset containing results from Polish Speedway League
 events \code{plsquads}
}
