% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointsCalc}
\alias{pointsCalc}
\title{Calculate points}
\usage{
pointsCalc(event_id, team_name, score)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Calculates points of \code{event_id} participant with given parameters
}

