% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{Ranking for multiple input}
\usage{
ranking(..., ties = "none")
}
\arguments{
\item{ties}{method of dealing with ties. Available 'first' and default='none'}

\item{a}{vector of attributes with prefix \code{a*} as variable to be ranked. Multiple arguments can be provided like \code{a1}, \code{a2}, ... \code{a10}. All arguments will be examined in given order to obtain ranking. If some observation have the same attribute (tie) then next argument will be decisive about ranking etc.}

\item{b}{vector of attributes with prefix \code{b*}}

\item{x}{vector with non-NA value pointing for an excluded from ranking. Ranking for \code{!is.na(x)} is being set to \code{NA} and others are updated}
}
\value{
A list with the elements
\item{y}{The sum of the squared values.}
}
\description{
Ranks multiple inputs with custom options
}
\note{
This is a very simple function.
}
\examples{
df <- 
  data.frame(
    group = c(rep("A",4),rep("B",4)),
    pts   = c(60,59,40,30,30,30,20,20),
    bal   = c(25,30,20,16,12,12,10,8),
    semi  = c('A','B','A','B',rep(NA,4)),
    final = c(1,NA,-2,rep(NA,5)),
    rel   = c(rep(NA,6),1,0),
    x     = c(rep(NA,3),1,rep(NA,4))
  )

df \%<>\% 
  mutate(
    ranking = ranking(a1=final, a2=semi, a3=pts, a4=bal, b1=rel, x1=x)
  ) \%>\% 
  arrange(ranking)
}

