% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{running_mean}
\alias{running_mean}
\title{Streak length of vector elements}
\usage{
running_mean(vec)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Calculates series of consecutive elements
}
\examples{
x <- c(1,1,0,0,1,0,3,3,1,1)
streakLength(x)
}
