% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.sport}
\alias{summary.sport}
\title{Summarizing sport objects}
\usage{
\method{summary}{sport}(object, ...)
}
\arguments{
\item{object}{of class sport}

\item{...}{optional arguments}
}
\description{
Summarizing sport objects
Summary for object of class `sport`
}
\examples{
model <- glicko_run(rank|id~rider, gpheats[1:100,]) 
summary(model)
}
