% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbt}
\alias{bbt}
\title{Bayesian Bradley-Terry model for single game}
\usage{
bbt(rank, mi_ij, sig_ij, kappa = 1e-04, gamma = 1)
}
\arguments{
\item{mi_ij}{ratings of participants.}

\item{sig_ij}{rating deviations of participants.}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates Glicko ratings based on Bayesian Bradley Terry model.
}
\details{
Algorithm based on 'A Bayesian Approximation Method for Online Ranking' by Ruby C. Weng and Chih-Jen Lin
}
