% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbt}
\alias{bbt}
\title{Bayesian Bradley-Terry model for single game}
\usage{
bbt(team_name, rank, r, rd, sig, weight, kappa = 1e-04, gamma = 1,
  init_r = 25, init_rd = 25/3, beta = 25/6)
}
\arguments{
\item{r}{ratings of participants.}

\item{rd}{rating deviations of participants.}

\item{beta}{}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates Glicko ratings based on Bayesian Bradley Terry model.
}
\details{
Algorithm based on 'A Bayesian Approximation Method for Online Ranking' by Ruby C. Weng and Chih-Jen Lin
}
