% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped_bbt.R
\name{bbt_run}
\alias{bbt_run}
\title{BBT rating algorithm}
\usage{
bbt_run(formula, data, r, rd)
}
\arguments{
\item{formula}{formula specifying model. BBT algorithm allows only player ranking parameter.}

\item{data}{data.frame which contains three columns: #' \enumerate{
  \item id event identifier
  \item team_name label of event participant -  name corresponding with parameters names
  \item rank position of team_name in event
}}

\item{r}{named vector of initial rating estimates. In there is no assumption, initial ratings should be r=25 Names of vector should correspond with team_name label.}

\item{rd}{named vector of initial rating deviation estimates. In there is no assumption, initial ratings should be r=25/3 Names of vector should correspond with team_name label.}
}
\description{
BBT rating algorithm
Wrapper arround `bbt` update algorithm. Wrapper allows user to simplify calculation providing only data and initial parameters assumptions
}
