% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbl}
\alias{dbl}
\title{Dynamic Bayesian Logit}
\usage{
dbl(team_name, rank, X, R, RD)
}
\arguments{
\item{X}{Matrix of player specifics.}

\item{R}{Matrix of coefficients (ratings).}

\item{RD}{Matrix of coefficients deviations}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates ratings using extended Kalman Filter.
}
