% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_dlr.R
\name{dbl_run}
\alias{dbl_run}
\title{DBL rating algorithm}
\usage{
dbl_run(formula, data, init_r, init_rd)
}
\arguments{
\item{formula}{formula specifying model.}

\item{data}{data.frame which contains three columns: #' \enumerate{
  \item id event identifier
  \item team_name label of event participant -  name corresponding with parameters names
  \item rank position of team_name in event
}}

\item{init_r}{named vector of initial rating estimates. In there is no assumption, initial ratings should be r=25 Names of vector should correspond with team_name label.}

\item{init_rd}{named vector of initial rating deviation estimates. In there is no assumption, initial ratings should be r=25/3 Names of vector should correspond with team_name label.}
}
\description{
DBL rating algorithm
Wrapper arround `dbl` update algorithm. Wrapper allows user to simplify calculation providing only data and initial parameters assumptions
}
