% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_run.R
\name{dbl_run}
\alias{dbl_run}
\title{Dynamic Bayesian Logit}
\usage{
dbl_run(
  formula,
  data,
  r = NULL,
  rd = NULL,
  lambda = NULL,
  weight = NULL,
  kappa = 0.95,
  init_r = 0,
  init_rd = 1
)
}
\arguments{
\item{formula}{formula which specifies the model. Unlike other algorithms
in the packages (glicko_run, glicko2_run, bbt_run), this doesn't allow
players nested in teams with `team(player | team)` but instead let user
specify multiple parameters also in interaction with others.}

\item{data}{data.frame which contains columns specified in formula, and
optionaly columns defined by `lambda`, `weight`.}

\item{r}{named vector of initial players ratings estimates. If not specified 
then `r` will be created automatically for parameters specified in `formula`
with initial value `init_r`.}

\item{rd}{rd named vector of initial rating deviation estimates. If not specified 
then `rd` will be created automatically for parameters specified in `formula`
with initial value `init_rd`.}

\item{lambda}{name of the column in `data` containing lambda values or one 
constant value (eg. `lambda = colname` or `lambda = 0.5`).
Lambda impact prior variance, and uncertainty of the matchup result. The 
higher lambda, the higher prior variance and more uncertain result of the 
matchup. Higher lambda flattens chances of winning.}

\item{weight}{name of the column in `data` containing weights values or
one constant (eg. `weight = colname` or `weight = 0.5`). 
Weights increasing (weight > 1) or decreasing (weight < 1) update change. 
Higher weight increasing impact of event result on rating estimate.}

\item{kappa}{controls `rd` shrinkage not to be greater than `rd*(1 - kappa)`.
`kappa=1` means that `rd` will not be decreased.}

\item{init_r}{initial values for `r` if not provided. 
Default (`glicko = 1500`, `glicko2 = 1500`, `bbt = 25`, `dbl = 0`)}

\item{init_rd}{initial values for `rd` if not provided. 
Default (`glicko = 350`, `glicko2 = 350`, `bbt = 25/3`, `dbl = 1`)}
}
\value{
A "rating" object is returned: \itemize{

\item \code{final_r} named vector containing players ratings.

\item \code{final_rd} named vector containing players ratings deviations.

\item \code{r} data.frame with evolution of the ratings and ratings deviations
 estimated at each event.
 
\item \code{pairs} pairwise combinations of players in analysed events with 
prior probability and result of a challange.

\item \code{class} of the object.

\item \code{method} type of algorithm used.

\item \code{settings} arguments specified in function call.
}
}
\description{
Dynamic Bayesian Logit
}
\examples{
# the simplest example
data <- data.frame(
  name = c("A", "B", "C", "D"),
  rank = c(3, 4, 1, 2),
  gate = c(1, 2, 3, 4),
  factor1 = c("a", "a", "b", "b"),
  factor2 = c("a", "b", "a", "b")
)

dbl <- dbl_run(
  data = data, 
  formula = rank ~ name
 )

dbl <- dbl_run(
  data = data, 
  formula = rank ~ team(name) + gate * factor1)
}
