% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl_run.R
\name{dbl_run}
\alias{dbl_run}
\title{DBL rating algorithm}
\usage{
dbl_run(formula, data, r, rd, sig, weight, date)
}
\arguments{
\item{formula}{formula specifying model. DBL allows multiple variables in formula, also two-way interaction are available. LHS needs `rank|id`, to specify competitors order and event `id`.}

\item{data}{data.frame which contains columns specified in formula, and optionaly columns defined by `sig`, `weight` or `date`.}

\item{r}{named vector of initial estimates. If there is no assumption, initial ratings is set to be r=0.}

\item{rd}{named vector of initial deviation estimates. In there is no assumption, initial is set to be rd=3.}

\item{sig}{named vector of rating volatile. In there is no assumption, initial ratings should be sig=0.5. Names of vector should correspond with team_name label.}

\item{weight}{name of column in `data` containing weights. Weights increasing or decreasing update change. Higher weight increasing impact of corresponding event.}

\item{date}{name of column in `data` containing date. Doesn't affect estimation process. If specified, charts displays estimates changes in time in}
}
\value{
A "sport" object is returned
\itemize{
  \item \code{final_r} named vector containing players ratings.
  \item \code{final_rd} named vector containing players ratings deviations.
  \item \code{r} data.frame with evolution of the ratings and ratings deviations estimated at each event.
  \item \code{pairs} pairwise combinations of players in analysed events with prior probability and result of a challange.
  \item \code{class} of the object
  \item \code{method} type of algorithm used
  \item \code{formula} modelled formula
}
}
\description{
DBL rating algorithm
Wrapper arround `dbl` update algorithm. Wrapper allows user to simplify calculation providing only data and initial parameters assumptions
}
