% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glicko}
\alias{glicko}
\title{Glicko}
\usage{
glicko(
  id,
  rank,
  name,
  r,
  rd,
  sigma,
  weight,
  init_r = 1500,
  init_rd = 350,
  gamma = 1,
  kappa = 0.5
)
}
\description{
Glicko
}
\examples{
x <- data.frame(id = 0, name = "a", rank = 1, r = 2, rd = 3, sigma = 4, weight = 5, identifier = 6)
glicko(
  id = x[["id"]], 
  name = x[["name"]], 
  rank = x[["rank"]], 
  r = x[["rank"]], 
  rd = x[["rd"]], 
  sigma = x[["sigma"]], 
  weight = x[["weight"]], 
  identifier = x[["identifier"]])
}
