% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glicko}
\alias{glicko}
\title{Glicko rating for single game}
\usage{
glicko(team_name, rank, r, rd, sig, weight, init_r = 1500, init_rd = 350,
  gamma = 1)
}
\arguments{
\item{team_name}{name of event participants.}

\item{rank}{classification of the event.}

\item{r}{ratings of participants.}

\item{rd}{rating deviations of participants.}

\item{sig}{}

\item{weight}{increase/decrease update of the parameter in particular event. Lower values makes parameter update smaller}

\item{init_r}{initial rating for new competitors (contains NA). Default = 1500}

\item{init_rd}{initial rating deviations for new competitors. Default = 350}

\item{time}{time after previous match - indicator multiplying uncertainty of expectations.}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates Glicko rating for single game input
}
\examples{
glicko(
 team_name = c( "A", "B", "C", "D" ), 
 rank  = c( 3, 4, 1, 2 ), 
 time  = c( 0, 0, 0, 0),
 r     = c( 1500, 1400, 1550, 1700 ) , 
 rd    = c( 200,  30,   100,  300 ),
 init_r  = 1500,
 init_rd = 100
)
glicko(
team_name = c( "A", "B" ), 
  rank  = c( 1, 2 ), 
  time  = c( 0, 0),
  r     = c( 1400, 1500 ) , 
  rd    = c( 80,  150),
  init_r  = 1500,
  init_rd = 100
)
}
