% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glicko}
\alias{glicko}
\title{Glicko rating for single game}
\usage{
glicko(name, rank, r, rd, sig, weight, init_r = 1500, init_rd = 350,
  gamma = 1)
}
\arguments{
\item{name}{of a player.}

\item{rank}{classification of the event.}

\item{r}{ratings of participants.}

\item{rd}{rating deviations of participants.}

\item{sig}{rating volitality is a value which multiplies prior `rd`. If `sig > 1` then prior `rd` increases, making estimate of `r` more uncertain.}

\item{weight}{increase/decrease update of the parameter in particular event. Lower values makes parameter update smaller}

\item{init_r}{initial rating for new competitors (contains NA). Default = 1500}

\item{init_rd}{initial rating deviations for new competitors. Default = 350}

\item{gamma}{can help to control how fast the variance `rd` is reduced after updating. Lower `gamma` slow down decreasing of `rd`, which tends to reach zero to quickly. The default value is `gamma = rd/c`.}
}
\value{
\code{r} updated ratings of participats

\code{rd} updated deviations of participants ratings

\code{expected} matrix of expected score. \code{expected[i, j] = P(i > j)}
}
\description{
Calculates Glicko rating for single game input
}
\examples{
glicko(
 name = c( "A", "B", "C", "D" ), 
 rank  = c( 3, 4, 1, 2 ), 
 r     = c( 1500, 1400, 1550, 1700 ) , 
 rd    = c( 200,  30,   100,  300 ),
 sig   = c(1,1,1,1),
 weight= c(1,1,1,1)
)
}
