% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glicko2_run.R
\name{glicko2_run}
\alias{glicko2_run}
\title{Glicko2 rating algorithm}
\usage{
glicko2_run(formula, data, r, rd, sig, weight)
}
\arguments{
\item{formula}{formula specifying model. Glicko2 algorithm allows only player ranking parameter and should be specified by following manner: 
`rank | id ~ name`. Names in formula are unrestricted, but model structure remains the same where:
\enumerate{
 \item rank alternative/participant position in experiment/event. Order of comparison
 \item id event identifier
 \item name of alternative/participant
}}

\item{data}{data.frame which contains columns specified in formula, and optionaly columns defined by `weight` or `date`.}

\item{r}{named vector of initial rating estimates. In there is no assumption, initial ratings should be r=1500. Names of vector should correspond with team_name label.}

\item{rd}{named vector of initial rating deviation estimates. In there is no assumption, initial ratings should be r=300 Names of vector should correspond with team_name label.}

\item{sig}{named vector of rating volatile. In there is no assumption, initial ratings should be sig=0.5. Names of vector should correspond with team_name label.}

\item{weight}{name of column in `data` containing weights. Weights multiplies step update increasing/decreasing step impact on parameters estimates}

\item{date}{name of column in `data` containing date. Doesn't affect estimation process. If specified, charts displays estimates changes in time instead of by observation `id`}
}
\description{
Glicko2 rating algorithm
Wrapper arround `glicko2` update algorithm. Wrapper allows user to simplify calculation providing only data and initial parameters assumptions
}
