% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl_wrapper.R
\name{glicko2_run}
\alias{glicko2_run}
\title{Glicko2 rating algorithm}
\usage{
glicko2_run(formula, data, r, rd, sig, time)
}
\arguments{
\item{data}{data.frame which contains three columns: #' \enumerate{
  \item id event identifier
  \item team_name label of event participant -  name corresponding with parameters names
  \item rank position of team_name in event
  \item time additional parameter increasing variance of participant rating related with time. 
}}

\item{r}{named vector of initial rating estimates. In there is no assumption, initial ratings should be r=1500. Names of vector should correspond with team_name label.}

\item{rd}{named vector of initial rating deviation estimates. In there is no assumption, initial ratings should be r=300. Names of vector should correspond with team_name label.}

\item{sig}{named vector of rating volatile. In there is no assumption, initial ratings should be sig=0.5. Names of vector should correspond with team_name label.}
}
\description{
Glicko2 rating algorithm
Wrapper arround `glicko2` update algorithm. Wrapper allows user to simplify calculation providing only data and initial parameters assumptions
}
