% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gpheats}
\alias{gpheats}
\title{Heat results of Speedway Grand-Prix}
\format{A data frame with >19000 rows and 11 variables:
\describe{
  \item{id}{event identifier}
  \item{season}{year of Grand-Prix, 1995-now}
  \item{date}{date of turnament}
  \item{round}{round in season}
  \item{name}{Turnament name}
  \item{heat}{heat number, 1-23}
  \item{field}{number of gate, 1-4}
  \item{rider}{rider name, string}
  \item{points}{paints gained, integer}
  \item{position}{position at finish line, string}
  \item{rank}{rank at finish line, integer}
}}
\source{
internal
}
\description{
Actual dataset containing heats results of all Speedway Grand-Prix turnaments
\code{gpheats}.
}
