% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gpsquads}
\alias{gpsquads}
\title{Turnament results of Speedway Grand-Prix}
\format{
A data frame with >4000 rows and 9 variables:
\describe{
\itemize{
\item \code{id} - event identifier
\item \code{season} - year of Grand-Prix, 1995-now
\item \code{date} - date of turnament
\item \code{place} - stadium of event
\item \code{round} - round in season
\item \code{name} - Turnament name
\item \code{rider} - rider names, 1-6
\item \code{points} - points gained, integer
\item \code{classification} - classification after an event
}
}
}
\source{
internal
}
\description{
Actual dataset containing turnament results of all Speedway Grand-Prix events
\code{gpsquads}
}
