% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gpsquads}
\alias{gpsquads}
\title{Turnament results of Speedway Grand-Prix}
\format{A data frame with >4000 rows and 9 variables:
\describe{
  \item{id}{event identifier}
  \item{season}{year of Grand-Prix, 1995-now}
  \item{date}{date of turnament}
  \item{place}{stadium of event}
  \item{round}{round in season}
  \item{name}{Turnament name}
  \item{rider}{rider names, 1-6}
  \item{points}{points gained, integer}
  \item{classification}{classification after an event}
}}
\source{
internal
}
\description{
Actual dataset containing turnament results of all Speedway Grand-Prix events
\code{gpsquads}
}
