% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot.rating}
\alias{plot.rating}
\title{Plot rating object}
\usage{
\method{plot}{rating}(x, n = 10, players, ...)
}
\arguments{
\item{x}{of class rating}

\item{n}{number of teams to be plotted}

\item{players}{optional vector with names of the contestants (coefficients) to plot 
their evolution in time.}

\item{...}{optional arguments}
}
\description{
Plot rating object
}
