% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rating}
\alias{predict.rating}
\title{Predict rating model}
\usage{
\method{predict}{rating}(object, newdata, ...)
}
\arguments{
\item{object}{of class rating}

\item{newdata}{data.frame with data to predict}

\item{...}{optional arguments}
}
\value{
probabilities of winning challenge by player over his opponent in all provided events.
}
\description{
Predict rating model
}
\examples{
glicko <- glicko_run(data = gpheats[1:16, ], 
                     formula = rank | id ~ player(rider))
predict(glicko, gpheats[17:20, ])
}
