% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sport}
\alias{predict.sport}
\title{Predict sport model}
\usage{
\method{predict}{sport}(object, new_data)
}
\arguments{
\item{object}{of class sport}

\item{new_data}{data.frame with data to predict}
}
\value{
probabilities of winning challange by player over his opponent in all provided events.
}
\description{
Predict sport model
}
\examples{
glicko <- glicko_run(rank|id~rider, gpheats[1:16,])
predict(glicko,gpheats[17:20,])
}
